Habitually Successful App
Created Habit tracking app with custom server

Summary:  App makes a call to the backend server to gather user data for login.  Once logged in user will have the ability to select
created habits or create new ones to mark as complete updating the days complete counter.  If a habit is not longer needed to be tracked user can delete a habit.  Sign out is available as well as sign up to create new users making calls to custom server with relational database deployed on heroku server

Screenshot: